Fix: allow turn off attributes in XML highlighting files
Summary:
In syntax highlighting XML files, it doesn't work to turn off attributes, for example: `bold="0"`, `italic="0"` or `underline="0"`.

This is fixed, but only for KSyntaxHighlighting themes, that is, this continues to not work in "KDE" and "Vim (dark)".

To correct that it is necessary to add to the API of the `KSyntaxHighlighting::Format` class, methods such as `hasBold()`, `hasItalic()`, etc. to know if the attributes "bold", "italic", "underline" & "strikeout" are modified in the XML file (the methods `isBold(...)`, `isItalic(...)`, etc., return the combination between the syntax definition and the theme).

Reviewers: #ktexteditor, dhaumann, cullmann

Reviewed By: #ktexteditor, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D23422