KateModeMenuList: add "Best Search Matches" section and fixes for Windows
Summary:
In the syntax highlighting menu, as the items are divided into sections, sometimes the best search match isn't shown first. For example, when writing "c", (in some translations) you have to navigate through the results to select the "C" language.

The **Best Search Matches** section is added, which shows the best matches at the top of the list. This search section appears only if the exact search matches aren't shown first. For example, when searching for the "C" or "R" language.

{F7349745}
{F7349749}

I also correct some bugs I saw in Kate for Windows:
* Don't overlap the scroll bar on the list in Windows, so that the selection of items looks better.
* The section separator line is semi-transparent in Windows, and the margins are smaller.
* Avoid always selecting section names.

Reviewers: #ktexteditor, cullmann, dhaumann

Reviewed By: #ktexteditor, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D23935