Add MarkInterfaceV2, to s/QPixmap/QIcon/g for symbols of marks
Summary:
Symbols used for marks can be used in different sizes, e.g. depending of the
line height, to which the icon border adapts, or for the context menu on
actions to toggle those marks. Being limited to set a single pixmap as
symbol for a mark results can result in badly scaled symbols being
displayed.
Switching to QIcon as dynamic pixmap provider for markers improves this.

For backward compatibility QIcon & QPixmaps are converted into each other in
case APIs are used mixed.

Reviewers: #kate, #kdevelop, dhaumann, cullmann

Reviewed By: #kate, dhaumann, cullmann

Subscribers: cullmann, anthonyfieroni, dhaumann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D27533