Fix modified line marker in kate minimap
Summary:
Our pixmap is the size of the number of lines in the document. When
rendering the pixel for the modified marker we want to take a ratio of
currentline to lines in the pixmap.

The height of the groove isn't relevant.

Test Plan:
Create a 10000 line empty document
Added some line in the middle
New line and orange dot appeared in the same place

Reviewers: #kate

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D29680