scrollbar minimap: performance: delay update for inactive documents
The scrollbar minimap would always be updated for documents when they changed,
even if the document was not visible at the time. In some situations, this
caused significant hangs in KDevelop (lots of documents open, then highlighting
being re-applied to all of them because of some background parser update e.g.).

Instead, we now remember that an update is required, but only do it as soon
as the document is shown.