[Vimode]Fix search inside fold ranges
Jumping to search matches in vimode (n/N) are implemented as motions and they are handled by NormalViMode::handleKeyPress which excludes code folds when jumping to lines. This works fine for most motions (e.g. jkG should skip folds to be consistent with Normal Kate mode), but introduces problems when matching searches as the cursor will be placed in position without the range being unfolded.

BUG: 376934