Correct indentation bug when line contains "for" or "else".
This commit replaces and extends 8a8c98ab.
Indeed, the root cause of the misindentation in 8a8c98ab is because cstyle.js
regex matches "for" and "else" even within words.
So, the following code:
if (true)
    fork(); // fork contains "for".
printf("PID: %d\n", getpid());
is indented as:
if (true)
    fork(); // fork contains "for".
    printf("PID: %d\n", getpid());
Now, this snippet is correctly indented and gives:
if (true)
    fork(); // fork contains "for".
printf("PID: %d\n", getpid());

Two tests were added to ensure the modifications are correct.