[KateViewInternal] Ensure that findMatchingBracket always works
For e.g. the text "{|}" (where | is the cursor), m_bmStart is
equal to [ (0, 0)  ->  (0, 1) ] and the closing bracket is in (0, 1).
Since m_bmStart->toRange().contains(m_cursor) is checked first, the
current bracket is incorrectly set to the opening bracket "{".

Mainly the vimode's % command (go to matching bracket) is affected.