Search: Enable Unicode support in QRegularExpression
I made a mistake by not enabling QRegularExpression::UseUnicodePropertiesOption
when porting QRegExp to QRegularExpression, it's not needed everywhere
but for a search tool that processes user input, it is essential to be
enabled. Note that QRegExp always enabled Unicode support by the default,
i.e. it wasn't an option AFAICS.

Extend the unit test by peppering it with some unicode characters.

Change and extend the "end case conversion", \E, unit test.

Thanks to Christoph for figuring that out.

See: https://invent.kde.org/frameworks/ktexteditor/-/issues/10