[Vimode] Improve sentence text object
Sentences starting on the first column were previously
not taken into account, so the entirety of:

This is a sentence.
This is another sentence.

only counted as a single sentence.

This patch also properly sets the starting position in visual mode.
The behavior is now closer to vim's sentence text objects but not
exactly the same. For example (| represents the start/end of selection)

This is a sentence. |  This is another sentence.

After the command "vis" in vim:

This is a sentence.|   |This is another sentence.

In Kate:
This is a sentence. |  This is another sentence.|