Fix incremental selection-only search
Since search matches are selected after they are found,
the "Search in selection only" option only works the first time.

This patch saves the initial selection where the search is initiated
and use that for succeeding selection-only searches. The selection
is "forgotten" only after the selection is manually changed or
the "selection only" option is toggled.

BUG: 327408