Use QPalette::highlight for the scrollbar minimap slider
Previously, QPalette::link() was used which is the color for hyperlinks.
Most themes set this to a shade of blue, so the slider will almost
always be blue even when, for example, the main color
of a theme is green.

The selection highlight color was chosen because it is also the color
used for indicating the active tab and is, generally, often attuned
to a theme's main color.

BUG: 434690