Use a MovingRange for remembering the selection search range
A MovingRange is needed since the range can change
when text is removed/replaced.

Currently, deleting a part of the selection search range,
then undoing that deletion will sometimes not properly
add that deleted range to the selection range (so the
search range will be effectively cut off). Fixing that
might need more changes in a separate patch.