Fix dragging when folding stuff is around
This will likely fix 435500 as well.

We were using selection cursors as if they were virtual cursors, which
lead to incorrect values and cases where "end < start". This will be
fixed now.

Pixmap position was incorrect, it would appear below the screen and not
attached to the "mouse cursor" as it should. This should also be fixed
now.

BUG: 435500