Add "Transpose Words" feature
Similar to "Transpose Characters", the word at the current cursor
position is swapped with the next word (or the previous word if
there is no next word in the line). Since KateView::word[Left|Right]
is used, the next word would be the word to the right for LTR
layouts and the word to the left for RTL layouts. The cursor's
relative position inside the word is also preserved.

For example, before transpose:
wor|d1  word2

after transpose:
word2 wor|d1

BUG: 436621