Fix on-the-fly spell checking with recent Qt
Qt commit 3a273ac47f20e82a1f2f63411b210025ca0f4495 changed the behavior of lastIndexOf, so it now starts matching the zero-length regexps from the end of the string.

This made the regexp '\b' match the end of the string, making the spell checker think that every character is a word by itself.

Fix it by ignoring the last character of the string when searching for the word boundary.

BUG: 436004