Ensure first column has a fixed size always
First column is the one that contains "arrow" when you have an expandable
item, i.e., an item that contains some documentation. If it is not expandable,
there is no arrow for it, this is okay however if no items in the view
are expandable, we get weird glitches and resizes sometimes.

To fix this, just leave space for it always.

Signed-off-by: Waqar Ahmed <waqar.17a@gmail.com>