Only process BOM for valid encodings (UTF-8/16/32)
Previously, if e.g. a ISO-8859-15 encoded file coincidentally
contains the UTF-8 BOM at the start, those three characters are
eaten and stripped from the text even though the Latin-15 encoding
does not have a BOM.

This commit makes it so that the encoding obtained from the BOM
is only used when it matches the previously set codec or when
no codec was set in the first place.