Dont indent the lines if line already has text + noindent
If you have a line with some text in it and you paste in the middle of
that line:

```
namespace
{
class SomeClass
```

Now imagine the cursor is after "class" and you paste "SOME_MACRO" there
the line gets reindented. I think it is better to preserve the existing
indentation in such a case.

Signed-off-by: Waqar Ahmed <waqar.17a@gmail.com>