KateWildcardMatcher: treat unmatched filename prefixes as no match
Extract the filename from the file path in
KateModeManager::wildcardsFind(), because otherwise the fix causes some
other tests to fail. Filenames, not entire paths should be matched
against wildcards.

This change fixes the following KateModeManagerTest::testWildcardsFind
data rows: "*qmakefile.cpp", "*_makefile.mm". The "not-Makefile.dic"
data row fails for a different reason now (negative file type priority).

The performance of KateModeManager::wildcardsFind() somewhat improved on
file paths and became somewhat worse on filenames without paths. Since
a full path to a file is usually passed to this function, it should be
faster on average now.

I have tried to replace the QFileInfo::fileName() call with a custom
fileNameFromFilePath() function, but the performance improvement was not
significant enough to warrant the code complexity increase.