The user is not selecting as soon as we clearSelection
This is something we only realised as soon as we started making more use
of this property. A way I found to track the regression in 6507add2 was
to use ctrl+backspace to remove a word. This would leave it in selecting
mode and the user would end up selecting even though shift or anything
is pressed.