Revert "Do not cancel mouse selection when using the keyboard"
This reverts commit 6507add26075ffd56e2efb4a35edcf059ec5d895.

reverted due to remaining issues in some corner cases

e.g.

There is an additional problem, not fixed yet. It goes away after reverting the changes from this merge request. Steps to reproduce:

open kwrite
write AAAA\nBBBB
navigate to the second line
press End
press and hold Shift
quickly tap Home and release Shift at as close to same time as possible - possibly one needs to release Shift a jiffy before releasing Home. Imagine bouncing/rotating the hand from the finger that presses Shift to the finger that presses Home.
when moving around with the cursor keys in the first line, it is clear that the editor is still in selection mode