Fix leaving selectionByUser state
Use keyPressEvent to check if shift (a modifier) is being pressed rather
than checking it in the release event. Otherwise there's the chance that
we'll be missing the modifier entirely if we do something like:
shift-press, end-press, shift-release, end-release.

In the example above the shift modifier would have never been registered
by a release event but the user sure had pressed it and expected it to
be taken into account.