Set metadata when creating katepart
Usually KPluginFactory takes care of this automatically, but given that
the creation process is non-standard here we need to do this manually.

This makes it possible to read the metadata defined in katepart.desktop
via part->metaData()

Unfortunatly KParts::Part::setMetaData is protected so we cannot access
it directly from KateFactory.

CCBUG: 444714