Apply word filter on async completion models
When an async model resets itself in a delayed fashion, we showed
all its items and ignored the word at the current cursor position.
By calling `cursorPositionChanged()` from the `modelReset()` slot
in `KateCompletionWidget` we can ensure that the newly arrived
results are properly filtered too.

This is important for KDevelop where the clang provided completion
items arrive with a noticeable delay. Something changed in handling
those items in the last few months, breaking the filtering. Now the
async results are correctly filtered again and the new unit test
coverage should hopefully prevent further breakage in the future.

BUG: 444883