Don't scroll on select all
Tried Qtcreator, browsers etc. No one jumps to end.

We use setSelection here to ensure we don't scroll anywhere
The cursor stays in place i.e., it doesn't move to end of selection
that is okay and expected.

The idea here is to maintain scroll position in case select all was
mistakenly triggered, and also to if you just want to copy text,
there is no need to scroll anywhere.

BUG: 435855