Reload doc on OnDiskCreated
Suppose you have two branches A and B. A has a file which B doesn't. If
you switch to B from A, while having that particular file opened, it will
be deleted. If you don't close the tab for it, and then switch back
to branch A, we miss to reload this file.