Remove superfluous QTextStream::setCodec call
QTextStream::setCodec doesn't do anything beyond updating the internal
codec pointer, and save/restore the seek position if there already is a
read buffer (which isn't the case here). Given our very next call is
setCodec() again, this call has no effect.

In itself this doesn't make much of a difference, but it helps with
reasoning about this code, which needs to be changed significantly for
Qt6.