Improve cursor movement with RTL text in doc
Currently, empty lines are always treated as LTR which is not a good
default to fall back on always. Instead if the line is empty, we can try
to scan nearby lines and choose a direction. This way it is possible for
RTL users to also navigate the text with keyboard.

BUG: 451789