Improve Color theme config page
Currently our color theme page is the most confusing config page of all.
There is theme editor and theme chooser both in one tab and it's often
really confusing which combobox will end up changing your editors color
scheme and which one belongs to the theme editor.

To fix this, move the theme editor into its own tab and create a new tab
for theme choosing. Also, add a KTE::View to the chooser page so that
the user can see in real time what the new theme will look like when
applied.