AppCommands: Do not quit application when last document or all documents are closed
This needs https://invent.kde.org/utilities/kate/-/merge_requests/758 in order to fix the bug.
Without it, we would simply ignore the "Close the application entirely when last file is closed"
setting if it is set to true. Though that is not mission-critical and IMHO less of a problem than the bug this commit fixes.

BUG: 454924