Set default mark type only if control is pressed
Currently the default mark type is always used, effectively making
it impossible to open the context menu.

This patch changes the behaviour in such a way that the default
mark type is only set if the control modifier is pressed. Now, if
control is not pressed (and more than one editable mark is set),
the context menu is shown.