Allow forcing RTL direction
The option is available under the "View" menu. Once the option is
checked, all the text is forced to the right. Having "Dynamic Word
Enabled" is necessary for this to work correctly. Reason being that
with static wrapping there is no "right" to which we can force the lines
to, "right" is infinity/INT_MAX.

BUG: 433926
BUG: 468151