Remove KateTextBlockTest
The stuff that this test covers is already covered by many other tests
since a block is a basic thing in KTextEditor. Moreover, it does the
tests in a different way from what the reality is i.e., creating raw
blocks with a dummy buffer. Its just bogus.

It also seems to be the only thing that needs Kate::TextRange exported.