Fix icon border not updated on save
Currently if you save a document which is modified, then the "red"
modified markers in the icon border stay red.

This used to work in KF5 but broke when I optimized the saveFile path
to not do anything in the readVariables() function if the file has no
variables. As a side effect of this optimization, the view wouldn't
repaint.

To fix this, trigger a repaint if:
- the view is visible and,
- it has modified lines in the visible range