InlineNote: Support RTL
InlineNotes currently don't support RTL at all which leads to a lot of
annoyance when using git-blame plugin.

This change adds support for inline-notes that are at the end of the
line.

The notes that are in the middle of line might still have issues but so
far we don't really have a usecase in which those notes appear in an
RTL context.