a11y: Specify buddies for "Appearance" -> "General"
Specify what the buddy for these labels from the
"Appearance" -> "General" tab in the settings is.

This allows to use the accelerator to jump to
the corresponding item and also makes sure the
labelled-by relationship gets set on the
accessibility layer, so the Orca screen reader no
longer just announces e.g. "checkbox checked" but
also the corresponding label, so it's clear what
the checkbox is actually for.