Optimize and simplify cursorAtBracket
Instead of laying out the lines, just multiply by spaceWidth. This is
what we do for normal indent lines so it makes sense to do it for the
highlighted indented lines.

Additionally, highlight () brackets as well as languages like
Dart/Flutter nest objects in () and a highlighted indent line can be
very helpful for navigation.

The range for finding bracket is increased to 500 lines as it is with
very large indents that line highlighting is actually helpful.