a11y: Improve tab order for "Appeareance" settings
Align the tab order (i.e. the order in which
widgets receive focus when moving through
the dialog using the Tab key) with the visual order
of the widgets in the "General" -> "Appearance"
settings page, so the focus doesn't unexpectedly
"jump" around.

To achieve that, move the widgets from row 2
of the layout above those from row 3 in the .ui file
as well, so they are constructed earlier. And drop the
manually specified tab order that was just including
two widgets.

This way, the widgets are constructed in the same
order as their visual order, and tabbing through
the page follows that order.

An alternative would be to manually specify the
tab order for all widgets manually, s. [1].

[1]  https://doc.qt.io/qt-6/designer-tab-order.html