a11y: Actually return line when requested
Previously, just the data for the word were returned
for the QAccessible::TextBoundaryType::LineBoundary case,
just like for QAccessible::TextBoundaryType::WordBoundary.

The QAccessible::TextBoundaryType::ParagraphBoundary case
already does what's needed, so use that for
QAccessible::TextBoundaryType::LineBoundary as well.

This makes the Orca screen reader announce the whole
line when moving between lines using the up/down arrow
key, instead of just annoucing the word.
(As a side note, the announcement is not 100% reliable,
but that's unrelated to this change.)