Fix selection shrink when indenting
'editRemoveText' indirectly triggers 'TextRange::checkValidity'
which silently shrink the selection if the removed text is at the start
or end of the given range.

'editInsertText' will insert text without expanding the selection.

The fix consists of doing insertion before deletion so that the
selection is not shrinked.

(more details in the Merge Request)

BUG: 329247
FIXED-IN:5.112