Introduce ArgumentHintWidget
This change replaces the old KateArgumentHintTree with
ArgumentHintWidget which is a simpler way to show the same data. The
problems with old widget were:
- It didn't work well with methods that have many overloads and long
or verbose signatures. There was too much visual noise
- It didn't work well with our new approach of not using top level
window for completion to avoid wayland bugs as the widget would get
cut off at the window edge
- It was too complicated code wise

The new widget is a simple plain text editor that shows the method
signature and allows switching between different available signatures
using Alt+Up/Down