a11y: Set "Line Height Multiplier" buddy
Specify the combobox as buddy of the
"Line Height Multiplier" label that's seen in Kate's
"Settings" -> "Configure Kate" -> "Appearance" -> "General"
tab.

This allows using the accelerator to jump to the
combobox and also makes sure the label-for/labelled-by
relationship get set on the accessibility layer,
so the Orca screen reader announces the label as well
when the combobox gets focus, so it's clear what this
is actually for.