a11y: Improve tab order for "Appeareance" -> "Borders"
Align the tab order (i.e. the order in which
widgets receive focus when moving through
the dialog using the Tab key) with the visual order
for the tab page shown in Kate's "Settings" ->
"Configure Kate" -> "Appearance" -> "Borders"
tab page, so the focus doesn't unexpectedly
jump around.

To achieve that, align the widget order in the .ui
file with the visual order on the tab page.

This way, the widgets are constructed in the same
order as their visual order, and tabbing through
the page follows that construction order.