Never remove trailing spaces on diff file type
So we have an option to remove trailing spaces on save, which is cool, except when you're manually editing patches (diff) and it messes up everything.

This patch checks if a file is of type Diff before applying the remove trailing spaces function.

There is no switch to turn it off in this implementation, though. But I can't think of a good reason to mess with trailing spaces in Diff.