use ktexteditor-script-tester6 for javascript tests
Apart from the format, the main changes are:

- systematic verification of cursor and selection positions
- addition of checks on intermediate steps for indentation tests

Indentation data files are moved to `indent2`. As a check is made on the
presence of files in the indentation data folders, this avoids errors when
switch to an older commit that restores them (switch to master does not delete
them as the `actual` files inside are not in the git tree).